refactor(animation): extract easing functions to separate module
Split animation.rs (1850 lines) into modular structure:

- easing.rs: All easing functions (linear, ease_in, ease_out, etc.)
- mod.rs: Animation, Tween, KeyframeAnimation, and other types

This improves code organization while maintaining the same API.

Closes #198 (partial)
